pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = [
  "setuptools>=67",
]

[project]
name = "drf-oidc-auth"
version = "4.0.0"
description = "Django Rest Framework OpenID Connect Authentication library"
readme = "README.md"
license = { text = "MIT License (MIT)" }
requires-python = ">=3.10"
classifiers = [
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "authlib>=1.0",
  "cryptography>=2.6",
  "django>=4.2",
  "djangorestframework>=3.14.0",
  "requests>=2.31.0",
]

[project.optional-dependencies]
allauth = [
  "django-allauth>=64.0",
]

[[project.authors]]
name = "Maarten van Schaik"
email = "support@byte.nl"

[project.urls]
Homepage = "https://github.com/ByteInternet/drf-oidc-auth"

## TESTING AND DEVELOPER CONFIGURATION BELOW ##
[tool.black]
line-length = 90
include = '\.pyi?$'

[tool.pyproject-fmt]
# after how many column width split arrays/dicts into multiple lines, 1 will force always
column_width = 120
# how many spaces use for indentation
indent = 2
# if false will remove unnecessary trailing ``.0``'s from version specifiers
keep_full_version = true
# maximum Python version to use when generating version specifiers
max_supported_python = "3.13"

[tool.coverage.run]
parallel = true
